Рефакторинг + комментарии в IoC
