+++ MathCore/IoC/IServiceManager.RegisterSingleton.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using MathCore.Annotations;
+using MathCore.IoC.ServiceRegistrations;
+
+namespace MathCore.IoC
+{
+    public partial interface IServiceManager
+    {
+        [NotNull] SingletonServiceRegistration<TService> RegisterSingleton<TService>() where TService : class;
+        [NotNull] SingletonServiceRegistration<TService> RegisterSingleton<TService>(TService ServiceInstance) where TService : class;
+        [NotNull] SingletonServiceRegistration<TService> RegisterSingleton<TServiceInterface, TService>() where TService : class, TServiceInterface;
+        [NotNull] SingletonServiceRegistration<TService> RegisterSingleton<TService>([NotNull] Func<TService> FactoryMethod) where TService : class;
+    }
+}
