+++ MathCore/IoC/ServiceManagerAccessor.cs
@@ -1,10 +1,12 @@
-﻿namespace MathCore.IoC
+﻿using MathCore.Annotations;
+
+namespace MathCore.IoC
 {
     public class ServiceManagerAccessor<TService> where TService : class
     {
         private readonly ServiceManager _ServiceManager;
 
-        public TService Service => _ServiceManager.Get<TService>();
+        [CanBeNull] public TService Service => _ServiceManager.Get<TService>();
 
         public ServiceManagerAccessor(ServiceManager ServiceManager) => _ServiceManager = ServiceManager;
     }
