+++ MathCore/IoC/ServiceRegistrations/SingleCallServiceRegistration.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 using MathCore.Annotations;
 
-namespace MathCore.IoC
+namespace MathCore.IoC.ServiceRegistrations
 {
     public class SingleCallServiceRegistration<TService> : ServiceRegistration<TService> where TService : class
     {
-        public SingleCallServiceRegistration(IServiceManager Manager, Type ServiceType) : base(Manager, ServiceType) { }
+        public SingleCallServiceRegistration(IServiceManager Manager, [NotNull] Type ServiceType) : base(Manager, ServiceType) { }
 
         public SingleCallServiceRegistration(IServiceManager Manager, Type ServiceType, Func<TService> FactoryMethod) : base(Manager, ServiceType, FactoryMethod) { }
 
@@ -17,11 +17,13 @@ namespace MathCore.IoC
             {
                 return CreateServiceInstance();
             }
+#pragma warning disable CA1031 // Do not catch general exception types
             catch (Exception e)
             {
                 LastException = e;
                 OnExceptionThrown(e);
             }
+#pragma warning restore CA1031 // Do not catch general exception types
 
             return null;
         }
